% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmf.simulate.R
\name{gmf.simulate}
\alias{gmf.simulate}
\title{Simulate a GLLVM model and draw a matrix from the model}
\usage{
gmf.simulate(n = 30, m = 60, d = 1, p = 2, family = poisson())
}
\arguments{
\item{n}{number of rows (observational units, subjects, etc.)}

\item{m}{number of columns (species, genes, etc.)}

\item{d}{number of fixed effects (observed)}

\item{p}{number of random effects (unobserved)}

\item{family}{a family as in the \code{\link{glm}} interface}
}
\value{
Returns a list with model parameters and simulated values
\itemize{
\item Y - simulated matrix of responses
\item X - simulated fixed effects
\item M - matrix of means \eqn{\mu_{ij}}
\item U - matrix of latent scores
\item V - matrix of latent variables
\item beta - matrix of coefficients of fixed effects
}
}
\description{
Generate model parameters of a GLLVM and draw a matrix from the model. Return the matrix, model parameters, and underlying
observed and unobserved factors. The drawn matrix is of size \eqn{n \times m}.
}
\details{
A matrix \eqn{[y_{ij}]} for \eqn{1 \leq i \leq n} and \eqn{1 \leq j \leq m} is drawn from
\eqn{y_{ij} | \mu_{ij} \sim \mathcal{F}(\mu_{ij},\phi_j)}
where \eqn{g} is a link function, i.e. \eqn{g(\mu_{ij}) = \eta_{ij} = \beta_{0j} + x_i'\beta_j + u_i'\lambda_j} for a
set distribution \eqn{\mathcal{F}}. Quantities \eqn{\beta}, \eqn{\lambda} and \eqn{\phi} are model parameters. \eqn{\beta}
correspond to fixed-effects or known factors. \eqn{\lambda} are random-effects or unobserved factors. \eqn{\phi} are dispersion parameters.

Model parameters as well as observed and unobserved parameters are simulated from a Gaussian distribution. These are tuned such that
\eqn{\eta_{ij}} are approximately on the scale \eqn{[-3,3]} enabling testing a wide variety of distributions and link functions.

To generate a matrix we first generate model parameters. Next, we generate observed \eqn{x_i} and latent scores \eqn{u_i}.
Together with the link function that identifies mean values \eqn{\mu_{ij}} which allow us to draw \eqn{y_{ij}}
from the defined distribution.
}
